Simplify surfer client functions
