+++ Cargo.toml
@@ -14,7 +14,5 @@ wit-bindgen = "0.36.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 bincode = "1.3.3"
-base64 = "0.22.0"
 leb128 = "0.2.5"
-eyre = "0.6.12"
