+++ src/webview/data_manager.ts
@@ -78,7 +78,7 @@ export class WaveformDataManager {
 
   addVariable(signalList: any, parentGroupId: number | undefined) {
     // Handle rendering a signal, e.g., render the signal based on message content
-    //console.log(message);
+    console.log("addVariable func", signalList);
 
     if (signalList.length === 0) {return;}
